Code style

+++ WindEnergy/Ext/TabControlExt.cs
@@ -1,16 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Drawing.Drawing2D;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using WindEnergy.WindLib.Classes.Collections;
-using WindEnergy.WindLib.Data;
 using WindEnergy.UI.Properties;
+using WindEnergy.WindLib.Classes.Collections;
 
 namespace WindEnergy.UI.Ext
 {
@@ -24,7 +17,7 @@ namespace WindEnergy.UI.Ext
             ShowToolTips = true;
             DrawMode = TabDrawMode.OwnerDrawFixed;
             this.ItemSize = new Size(100, 20);
-            this.Font = new Font("Microsoft Sans Serif", 9.75F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
+            this.Font = new Font("Microsoft Sans Serif", 9.75F, FontStyle.Regular, GraphicsUnit.Point, 0);
         }
 
         /// <summary>
@@ -61,7 +54,7 @@ namespace WindEnergy.UI.Ext
 
             //закрытие при нажатии СКМ
             if (e.Button == MouseButtons.Middle)
-                for (int i = 0; i < this.TabPages.Count;i++)
+                for (int i = 0; i < this.TabPages.Count; i++)
                 {
                     Rectangle rect = GetTabRect(i);
                     if (rect.Contains(pt))
@@ -116,6 +109,6 @@ namespace WindEnergy.UI.Ext
                 }
             }
         }
-        
+
     }
 }
